% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dge.R
\name{run_dge}
\alias{run_dge}
\title{Diferansiyel Gen Ekspresyonu Analizi}
\usage{
run_dge(expr_mat, groups)
}
\arguments{
\item{expr_mat}{Gen ekspresyon matrisi (log2 olmali)}

\item{groups}{Orneklerin gruplari (vektor)}
}
\value{
topTable sonucu
}
\description{
Diferansiyel Gen Ekspresyonu Analizi
}
