% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_voom_dge.R
\name{run_voom_dge}
\alias{run_voom_dge}
\title{RNA-seq Ham Sayim Verileri Icin limma-voom Analizi}
\usage{
run_voom_dge(counts_mat, group_labels)
}
\arguments{
\item{counts_mat}{Ham RNA-seq sayim matrisi (satirlar: genler, sutunlar: ornekler)}

\item{group_labels}{Orneklerin gruplari (vektor, orn: c('Control', 'Treatment'))}
}
\value{
topTable formatinda diferansiyel ekspresyon tablosu
}
\description{
RNA-seq Ham Sayim Verileri Icin limma-voom Analizi
}
